Set the focus control to null, so that when we switch items, there is no focus control
Set an option in the list view to grey out the panel instead of disabling the whole thing. This option is commented out

Added a weapons material list to the database. Now you can define a material for the weapon. Currently it is stored in an enum script. In the future, might want to make it a database as well...